fix(jpeg): Fix several issues reported recently,
1. Fix decode images with dri marker failed,
2. Fix encode sometimes get length error

+++ components/esp_driver_jpeg/jpeg_private.h
@@ -87,7 +87,7 @@ typedef struct {
     uint8_t huffcode[2][2][JPEG_HUFFMAN_AC_VALUE_TABLE_LEN];    // Huffman decoded data tables [id][dcac]
     uint32_t tmp_huff[JPEG_HUFFMAN_AC_VALUE_TABLE_LEN];         // temp buffer to store huffman code
     bool dri_marker;                                            // If we have dri marker in table
-    uint8_t ri;                                                 // Restart interval
+    uint16_t ri;                                                // Restart interval
 } jpeg_dec_header_info_t;
 
 struct jpeg_decoder_t {
